# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for managing Akun resources",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/data_akun": {
            "get": {
                "tags": [
                    "DataAkun"
                ],
                "summary": "Get list of DataAkun",
                "description": "Display a listing of DataAkun.",
                "operationId": "51a17f8a35fd8664b90191d6f4371a1d",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DataAkun"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "DataAkun"
                ],
                "summary": "Create a new DataAkun",
                "description": "Store a new DataAkun.",
                "operationId": "42f086e8a2eca04493b5dc285ce1930b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "id_sub_akun": {
                                        "type": "string",
                                        "example": "akun_123"
                                    },
                                    "debit": {
                                        "type": "integer",
                                        "example": "20000"
                                    },
                                    "kredit": {
                                        "type": "integer",
                                        "example": "20000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "DataAkun created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataAkun"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/data_akun/{id}": {
            "get": {
                "tags": [
                    "DataAkun"
                ],
                "summary": "Get a DataAkun by ID",
                "description": "Display a DataAkun by ID.",
                "operationId": "db1144ac819c716e1c8452a58865acf7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataAkun"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "DataAkun not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "DataAkun"
                ],
                "summary": "Update a DataAkun by ID",
                "description": "Update an existing DataAkun.",
                "operationId": "8775638f946c9d40adbb66bb025c32aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "id_sub_akun": {
                                        "type": "string",
                                        "example": "akun_456"
                                    },
                                    "debit": {
                                        "type": "integer",
                                        "example": "2000"
                                    },
                                    "kredit": {
                                        "type": "integer",
                                        "example": "2000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "DataAkun updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataAkun"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "DataAkun not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "DataAkun"
                ],
                "summary": "Delete a DataAkun by ID",
                "description": "Delete a DataAkun.",
                "operationId": "593119dff1d1fc49551c70ab44548adf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "DataAkun deleted"
                    },
                    "404": {
                        "description": "DataAkun not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DataAkun": {
                "properties": {
                    "id_data_akun": {
                        "type": "string",
                        "example": "sub_akun_123"
                    },
                    "id_sub_akun": {
                        "type": "string",
                        "example": "akun_123"
                    },
                    "kode": {
                        "type": "integer",
                        "example": 101
                    },
                    "nama": {
                        "type": "string",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}